# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/health":{"get":{"operationId":"AppController_health","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["App"]}},"/api/v1/users/me":{"get":{"operationId":"UsersController_me_v1","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"accessToken":[]}],"tags":["Users"]},"patch":{"operationId":"UsersController_updateMe_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"security":[{"accessToken":[]}],"tags":["Users"]}},"/api/v1/listings":{"post":{"operationId":"ListingsController_create_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateListingDto"}}}},"responses":{"201":{"description":""}},"security":[{"accessToken":[]}],"tags":["Listings"]},"get":{"operationId":"ListingsController_findAll_v1","parameters":[],"responses":{"200":{"description":""}},"tags":["Listings"]}},"/api/v1/listings/me":{"get":{"operationId":"ListingsController_findMyListings_v1","parameters":[],"responses":{"200":{"description":""}},"security":[{"accessToken":[]}],"tags":["Listings"]}},"/api/v1/listings/{id}":{"get":{"operationId":"ListingsController_findOne_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Listings"]},"patch":{"operationId":"ListingsController_update_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateListingDto"}}}},"responses":{"200":{"description":""}},"security":[{"accessToken":[]}],"tags":["Listings"]},"delete":{"operationId":"ListingsController_remove_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"accessToken":[]}],"tags":["Listings"]}},"/api/v1/offers":{"post":{"operationId":"OffersController_create_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOfferDto"}}}},"responses":{"201":{"description":""}},"tags":["Offers"]},"get":{"operationId":"OffersController_findAll_v1","parameters":[],"responses":{"200":{"description":""}},"tags":["Offers"]}},"/api/v1/offers/{id}":{"get":{"operationId":"OffersController_findOne_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Offers"]},"patch":{"operationId":"OffersController_update_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOfferDto"}}}},"responses":{"200":{"description":""}},"tags":["Offers"]},"delete":{"operationId":"OffersController_remove_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Offers"]}},"/api/v1/sessions":{"get":{"operationId":"SessionsController_findAll_v1","parameters":[],"responses":{"200":{"description":""}},"security":[{"accessToken":[]}],"tags":["Sessions"]},"delete":{"operationId":"SessionsController_revokeAll_v1","parameters":[],"responses":{"200":{"description":""}},"security":[{"accessToken":[]}],"tags":["Sessions"]}},"/api/v1/sessions/{id}":{"delete":{"operationId":"SessionsController_revoke_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"accessToken":[]}],"tags":["Sessions"]}},"/api/v1/auth/register":{"post":{"operationId":"AuthController_register_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/refreshToken":{"post":{"operationId":"AuthController_refresh_v1","parameters":[],"responses":{"200":{"description":""}},"security":[{"refresh-token":[]}],"tags":["Auth"]}},"/api/v1/auth/logout":{"post":{"operationId":"AuthController_logout_v1","parameters":[],"responses":{"200":{"description":""}},"security":[{"accessToken":[]}],"tags":["Auth"]}}},"info":{"title":"My API","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"UpdateProfileDto":{"type":"object","properties":{"firstName":{"type":"string","example":"John"},"lastName":{"type":"string","example":"Doe"}}},"UserEntity":{"type":"object","properties":{"id":{"type":"number"},"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"listingsCount":{"type":"number"},"offersCount":{"type":"number"}},"required":["id","email","firstName","lastName","createdAt","updatedAt"]},"CreateListingDto":{"type":"object","properties":{"name":{"type":"string","minLength":3,"example":"Luxury Beachfront Villa","description":"The display name of the property"},"description":{"type":"string","minLength":15,"example":"A stunning 3-bedroom villa with a private pool and sunset views over the Pacific.","description":"Detailed description (min 15 chars)"},"price":{"type":"number","minimum":1,"example":450,"description":"Price per night in USD"},"address":{"type":"string","example":"123 Ocean Drive, Miami, FL"},"type":{"type":"string","enum":["RESIDENTIAL","COMMERCIAL","BUSINESS"],"example":"BUSINESS"},"status":{"type":"string","enum":["DRAFT","PUBLISHED","SOLD"],"example":"PUBLISHED"},"attributes":{"type":"object","example":{"wifi":true,"parking":"2 spots","pool":"private"}}},"required":["name","description","price","type"]},"UpdateListingDto":{"type":"object","properties":{"name":{"type":"string","minLength":3,"example":"Luxury Beachfront Villa","description":"The display name of the property"},"description":{"type":"string","minLength":15,"example":"A stunning 3-bedroom villa with a private pool and sunset views over the Pacific.","description":"Detailed description (min 15 chars)"},"price":{"type":"number","minimum":1,"example":450,"description":"Price per night in USD"},"address":{"type":"string","example":"123 Ocean Drive, Miami, FL"},"type":{"type":"string","enum":["RESIDENTIAL","COMMERCIAL","BUSINESS"],"example":"BUSINESS"},"status":{"type":"string","enum":["DRAFT","PUBLISHED","SOLD"],"example":"PUBLISHED"},"attributes":{"type":"object","example":{"wifi":true,"parking":"2 spots","pool":"private"}}}},"CreateOfferDto":{"type":"object","properties":{"amount":{"type":"number"},"buyerId":{"type":"number"},"listingId":{"type":"number"}},"required":["amount","buyerId","listingId"]},"UpdateOfferDto":{"type":"object","properties":{"amount":{"type":"number"},"buyerId":{"type":"number"},"listingId":{"type":"number"}}},"RegisterDto":{"type":"object","properties":{"email":{"type":"string","format":"email","example":"user@example.com"},"password":{"type":"string","minLength":8,"maxLength":64,"example":"password123"},"firstName":{"type":"string","minLength":1,"maxLength":50,"example":"Yahya"},"lastName":{"type":"string","minLength":1,"maxLength":50,"example":"Sherbini"}},"required":["email","password","firstName","lastName"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","format":"email","example":"user@example.com"},"password":{"type":"string","example":"password123"}},"required":["email","password"]}}}}